Rename migrate redis data CLI file
