+++ lib/onetime/cli.rb
@@ -27,7 +27,7 @@ end
 # Load CLI commands
 require_relative 'cli/migrate_command'
 require_relative 'cli/change_email_command'
-require_relative 'cli/consolidate_redis_models_command'
+require_relative 'cli/migrate_redis_data_command'
 require_relative 'cli/customers_command'
 require_relative 'cli/domains_command'
