+++ lib/onetime/cli/migrate_redis_data_command.rb
